testAPI/ads/serializers.py
from rest_framework import serializers

from .models import Advertisement


class AdvertisementSerializer(serializers.ModelSerializer):
    class Meta:
        model = Advertisement
        fields = ["ad_id", "title", "author", "views", "position"]
