requirements.txt
Django
djangorestframework
django-rest-framework-simplejwt